backend/requirements.txt
### backend/requirements.txt
fastapi
uvicorn
requests
google-auth
google-auth-httplib2
google-api-python-client
